+++ crates/arch/src/lib.rs
@@ -5,9 +5,6 @@ pub mod cpuid;
 pub mod hcf;
 pub mod interrupt;
 pub mod msr;
-pub mod nmi;
-pub mod paging;
-pub mod rtc;
+
 pub mod speaker;
-pub mod tsc;
 pub use msr::rdmsr;
