+++ crates/bit-fade/Cargo.toml
@@ -5,6 +5,6 @@ edition = "2024"
 
 [dependencies]
 commons.workspace = true
-arch.workspace  = true
+timers.workspace = true
 limine.workspace = true
 boot.workspace = true
