+++ crates/paging/Cargo.toml
@@ -1,12 +1,9 @@
 [package]
-name = "arch"
+name = "paging"
 version = "0.1.0"
 edition = "2024"
 
 [dependencies]
 x86_64.workspace = true
-sync.workspace = true
-fb.workspace = true
 boot.workspace = true
 allocators.workspace = true
-timers.workspace = true
