+++ kernel/Cargo.toml
@@ -18,8 +18,9 @@ lazy_static = { version = "1.5.0", default-features = false, features = [
 ] }
 limine = "0.5.0"
 bit-fade.workspace = true
-
+timers.workspace = true
 x86_64.workspace = true
 embedded-graphics.workspace = true
 x2apic.workspace = true
 allocators.workspace = true
+paging.workspace = true
